* Changed the zdtm_env_t socket members from the unix standard integer type to the newly defined cross-platform type SOCKET.
* Modified the _zdtm_listen_for_zaurus() function in the zdtm_net.c file
to be cross-platform between Unix Varients and Windows.

* Modified the _zdtm_handle_zaurus_conn() function in the zdtm_net.c
file to be cross-platform between Unix Varients and Windows.

* Modified the _zdtm_close_zaurus_conn() function in the zdtm_net.c file
to be cross-platform between Unix Varients and Windows.

* Modified the _zdtm_conn_to_zaurus() function in the zdtm_net.c file to
be cross-platform between Unix Varients and Windows.

* Modified the _zdtm_close_conn_to_zaurus() function in the zdtm_net.c
file to be cross-platform between Unix Varients and Windows.

* Modified the _zdtm_send_comm_message_to() function in the zdtm_net.c
file and zdtm_net.h file to be cross-platform between Unix Varients and
Windows. I also rewrote the algorithm to always write COM_MSG_SIZE
bytes. It had a problem before where if due to buffer issues less than
COM_MSG_SIZE bytes were written then it would return in error.

* The _zdtm_send_ack_message(), _zdtm_send_rqst_message(), and
_zdtm_send_abrt_message() functions needed NO modification since they
are strictly interface functions to the _zdtm_send_comm_message_to()
function.

* Rewrote the _zdtm_recv_message() function in the zdtm_net.c file to be
cross-platform between Unix Varients and Windows. I also rewrote it from
scratch to make sure we are recving all the necessary/available bytes
given the that recv() does not gaurante one gets the number of bytes
they request.

* The _zdtm_send_message() function needed NO modification since it is
strictly an interface function to the _zdtm_send_message_to() function.

* Modified the _zdtm_send_message_to() function in the zdtm_net.c file
to be cross-platform between Unix Varients and Windows. I also rewrote a
portion of it so that it will write all the messages bytes, without
having to worry about buffering issues.

* Wrote the content for the README file to describe appropriately how to
build on Unix, Mac OS X, and Windows.

* Modified the configure.ac to handle the host specific build flags.

* Modified the src/Makefile.am to use the new host specific LDFLAGS
provided in the configure.ac.

* Did both Linux and Windows tests on Debian (Etch = Testing), all
builds with 0 errors, 0 warnings. However, the functionality has not
been tested.

* Removed testing directory from the root Makefile.am file so that the
testing directory wouldn't be built. The testing directory was custom
testing designed for unix version only, and hence uses functions that
are not exported via the windows DLL yet. Once, they are ported over
it will be re-inserted into the root Makefile.am.

+++ testing/zdtm_test_daemon.c
@@ -143,22 +143,22 @@ int test_get_changeinfo(zdtm_lib_env *cur_env) {
     printf("num new sync ids = %u\n", rmsg.body.cont.asy.num_new_sync_ids);
     for (i = 0; i < rmsg.body.cont.asy.num_new_sync_ids; i++) {
         printf(" -- new sync id [%d] = (hex) 0x%.8x, (dec) %u\n", i,
-            rmsg.body.cont.asy.new_sync_ids[i],
-            rmsg.body.cont.asy.new_sync_ids[i]);
+            (int)rmsg.body.cont.asy.new_sync_ids[i],
+            (int)rmsg.body.cont.asy.new_sync_ids[i]);
     }
     printf("mod list id = 0x%.2x\n", rmsg.body.cont.asy.mod_list_id);
     printf("num mod sync ids = %u\n", rmsg.body.cont.asy.num_mod_sync_ids);
     for (i = 0; i < rmsg.body.cont.asy.num_mod_sync_ids; i++) {
         printf(" -- mod sync id [%d] = (hex) 0x%.8x, (dec) %u\n", i,
-            rmsg.body.cont.asy.mod_sync_ids[i],
-            rmsg.body.cont.asy.mod_sync_ids[i]);
+            (int)rmsg.body.cont.asy.mod_sync_ids[i],
+            (int)rmsg.body.cont.asy.mod_sync_ids[i]);
     }
     printf("del list id = 0x%.2x\n", rmsg.body.cont.asy.del_list_id);
     printf("num del sync ids = %u\n", rmsg.body.cont.asy.num_del_sync_ids);
     for (i = 0; i < rmsg.body.cont.asy.num_del_sync_ids; i++) {
         printf(" -- del sync id [%d] = (hex) 0x%.8x, (dec) %u\n", i,
-            rmsg.body.cont.asy.del_sync_ids[i],
-            rmsg.body.cont.asy.del_sync_ids[i]);
+            (int)rmsg.body.cont.asy.del_sync_ids[i],
+            (int)rmsg.body.cont.asy.del_sync_ids[i]);
     }
 
     _zdtm_clean_message(&rmsg);
